% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\docType{data}
\name{census}
\alias{census}
\title{Chinese census population counts by age and sex}
\format{
An array with dimensions "age",
"sex", and "time".
}
\source{
Derived from table "Population by age, sex
and urban/rural residence"
in the United Nations Statistics Division online database
"Population Censuses Datasets (1995 - Present)", 
downloaded on 21 January 2018.
}
\usage{
census
}
\description{
Census data on population counts by age and sex,
for 1990, 2000, and 2010. The maximum age group is 80+.
}
\keyword{datasets}
