% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_marital_status.R
\docType{data}
\name{conc_marital_status}
\alias{conc_marital_status}
\title{Concordance between marital statuses}
\format{
An data.frame with columns "from" and "to".
}
\usage{
conc_marital_status
}
\description{
A concordance between the marital statuses
used in \code{\link{remarriages}} and the
marital statuses used
for all other datasets. The \code{\link{remarriages}}
dataset does not distinguish between "Divorced"
and "Widowed".
}
\keyword{datasets}
