% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths_cb.R
\docType{data}
\name{deaths_cb}
\alias{deaths_cb}
\title{Estimates of deaths in China from the US Census Bureau}
\format{An array with dimension "time".}
\source{
Calculated from data downloaded from the
US Census Bureau International Data Base
table "Components of Population Growth",
downloaded on 28 March 2019.
}
\usage{
deaths_cb
}
\description{
Total numbers of deaths in China, by 5-year period, as estimated
by the United States Census Bureau.
}
\keyword{datasets}
