% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths_un.R
\docType{data}
\name{deaths_un}
\alias{deaths_un}
\title{Estimates of deaths in China from the UN Population Division}
\format{
An array with dimensions "age",
"sex", and "time"
}
\source{
Table "Number of deaths by age and sex (thousands)"
in the United Nations Population Division online database
"World Population Prospects".
Data  downloaded on 2 December 2017.
}
\usage{
deaths_un
}
\description{
Total numbers of deaths in China, by age, sex, and 5-year period, as estimated
by the United Nations Population Division.
}
\keyword{datasets}
