% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeexp_un.R
\docType{data}
\name{lifeexp_un}
\alias{lifeexp_un}
\title{Estimates of life expectancy at birth in China from the UN Population Division}
\format{
An array with dimensions "sex", and "time"
}
\source{
Table "Life expectancy, e(x), at exact age x (years)"
in the United Nations Population Division online database
"World Population Prospects".
Data  downloaded on 27 July 2020.
}
\usage{
lifeexp_un
}
\description{
Life expectancy at birth, by sex, and 5-year period, as estimated
by the United Nations Population Division.
}
\keyword{datasets}
