% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bar.R
\name{make_bar}
\alias{make_bar}
\title{make_bar make a bar chart}
\usage{
make_bar(x, colors, title)
}
\arguments{
\item{x}{- a data.frame to plot}

\item{colors}{- a vector that decide the group colors}

\item{title}{- the title of the plot}
}
\description{
make_bar make a bar chart
}
\examples{
x<-matrix(c(500,2000,36),3,1,dimnames =list(c("a","b","c")))
          names(x)<-c("a","b","c")
           make_bar(x, c("red","blue","yellow"), "title")

}
