% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marital_status_1990.R
\docType{data}
\name{marital_status_1990}
\alias{marital_status_1990}
\title{UN estimates of Chinese population aged 15+ by marital status in 1990}
\format{An array with dimensions "status", "age",
"sex", and "time".}
\source{
Table "Population by marital status,
age, sex and urban/rural residence"
in the United Nations Statistics Division online database
"Demographic Statistics Database".
Data  downloaded on 14 June 2019.
}
\usage{
marital_status_1990
}
\description{
Chinese population aged 15+, by age, sex, and marital status
"Single", "Married", "Divorced", and "Widowed" in 1990. The data
are derived from the 1990 Chinese Census, as reported by the UN.
}
\details{
The maximum age group for this dataset is 60+.
}
\seealso{
\code{\link{marital_status_2000_2010}}
}
\keyword{datasets}
