% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marital_status_2000_2010.R
\docType{data}
\name{marital_status_2000_2010}
\alias{marital_status_2000_2010}
\title{UN estimates of Chinese population aged 15+ by marital status
in 2000 and 2010}
\format{An array with dimensions "status", "age",
"sex", and "time".}
\source{
Table "Population by marital status,
age, sex and urban/rural residence"
in the United Nations Statistics Division online database
"Demographic Statistics Database".
Data  downloaded on 14 June 2019.
}
\usage{
marital_status_2000_2010
}
\description{
Chinese population aged 15+, by age, sex, and marital status
"Single", "Married", "Divorced", and "Widowed" in 2000 and 2010.
The data are derived from the 2000 and 2010 Chinese Censuses,
as reported by the UN.
}
\details{
The maximum age group for this dataset is 65+.

The UN database reports the counts from the census long form,
which was given to 10% of respondents. We have multiplied
the UN counts by 10.

The 2000 census includes a category "In consensual union". The
number of people in this category is very small, and we have
put them in the "Married" category.
}
\seealso{
\code{\link{marital_status_1990}}
}
\keyword{datasets}
