% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marriages_divorces.R
\docType{data}
\name{marriages_divorces}
\alias{marriages_divorces}
\title{Numbers of marriages and divorces in China}
\format{An array with dimensions "status_orig", "status_dest",
and "time".}
\source{
JUNNI TO COMPLETE
}
\usage{
marriages_divorces
}
\description{
Numbers of movements between marriage statuses, calculated
from official Chinese data on marriages, divorces, and
remarriages.
}
\seealso{
\code{\link{remarriages}}
}
\keyword{datasets}
