% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popn_survey.R
\docType{data}
\name{popn_survey}
\alias{popn_survey}
\title{Counts from Population Sample Survey}
\format{An array with dimensions "sex", "status", and "time".}
\source{
Table 4.11 from the 2006 Chinese Statistical Yearbook,
and Table 2.13 from the 2016 Chinese Statistical Yearbook,
avaliable online at website for the National Bureau of
Statistics of China,
http://www.stats.gov.cn/english/Statisticaldata/AnnualData/.
Downloaded on 4 September 2019.
}
\usage{
popn_survey
}
\description{
Counts of population aged 15+ by sex and marital
status from 2005 and 2015 Population Sample Surveys.
The sampling fraction in 2005 was 1.325\% and the sampling
fraction in 2015 was 1.55\%.
}
\keyword{datasets}
