% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popn_survey_fraction.R
\docType{data}
\name{popn_survey_fraction}
\alias{popn_survey_fraction}
\title{Sampling fractions from Population Sample Survey}
\format{An array with dimension "time".}
\source{
Table 4.11 from the 2006 Chinese Statistical Yearbook,
and Table 2.13 from the 2016 Chinese Statistical Yearbook,
avaliable online at website for the National Bureau of
Statistics of China,
http://www.stats.gov.cn/english/Statisticaldata/AnnualData/.
Downloaded on 4 September 2019.
}
\usage{
popn_survey_fraction
}
\description{
The proportion of the total population aged 15+ that
was included in the 2005 and 2015 Population Sample Surveys.
}
\seealso{
\code{\link{popn_survey}}
}
\keyword{datasets}
