% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_cb.R
\docType{data}
\name{population_cb}
\alias{population_cb}
\title{Estimates of population of China from the US Census Bureau}
\format{An array with dimensions "age",
"sex", and "time"}
\source{
Calculated from data downloaded from the
US Census Bureau International Data Base
table "Mid-year Population by Five Year Age Groups and Sex",
downloaded on 28 March 2019.
}
\usage{
population_cb
}
\description{
Population of China by 5-year age group and sex, at 5-year
intervals, as estimated
by the United States Census Bureau.
}
\keyword{datasets}
