% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_un.R
\docType{data}
\name{population_un}
\alias{population_un}
\title{Census data on population of China from the UN}
\format{An array with dimensions "age",
"sex", and "time"}
\source{
Table "Population by age, sex and urban/rural residence"
in the United Nations Statistics Division online database
"Population Censuses Datasets (1995 - Present)".
Data  downloaded on 21 January 2018.
}
\usage{
population_un
}
\description{
Census counts for China by 5-year age group and sex, at 10-year
intervals, as reported by the United Nations.
}
\keyword{datasets}
