% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remarriages.R
\docType{data}
\name{remarriages}
\alias{remarriages}
\title{Numbers of remarriages in China}
\format{An array with dimensions "status_orig", "status_dest",
and "time".}
\source{
JUNNI TO COMPLETE
}
\usage{
remarriages
}
\description{
Numbers of movements between "Divorced" or "Widowed"
and "Married", calculated
from official Chinese data on marriages, divorces, and
remarriages. Note that, using the official data, it
is not possible to distinguish between transitions
from divorced to married and transitions from
widowed to married, so the dataset combines the two.
}
\seealso{
\code{\link{marriages_divorces}},
\code{\link{conc_marital_status}}
}
\keyword{datasets}
