% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agePlot.R
\name{agePlot}
\alias{agePlot}
\title{agePlot Helper function that draws a bar chart when passed a sex and marriage categories vector}
\usage{
agePlot(thePlace, theGroup, theSex)
}
\arguments{
\item{thePlace}{Vector that refers to a place}

\item{theGroup}{for marital status group which}

\item{theSex}{Females or Males}
}
\description{
agePlot Helper function that draws a bar chart when passed a sex and marriage categories vector
}
\examples{
agePlot("Miami-Dade County", "Never married", theSex = "Male")
agePlot("Alachua County", "Divorced", theSex = "Female")


}
