% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births_cb.R
\docType{data}
\name{births_cb}
\alias{births_cb}
\title{Estimates of births in China from the US Census Bureau}
\format{An array with dimension "time".}
\source{
Calculated from data downloaded from the
International Database (IDB) on 28 March 2019.
}
\usage{
births_cb
}
\description{
Total numbers of births in China, by 5-year period, as estimated
by the United States Census Bureau.
}
\keyword{datasets}
